controller that loses to randoms: (
